requirements.txt
pyserial
pyqtgraph==0.12.4
PyQt5==5.15.2
